Delete unused files and set all products to redux store
