+++ src/redux/action/actionCreators.js
@@ -1,5 +1,5 @@
 /* eslint-disable import/prefer-default-export */
-import { FETCH_CATEGORIES } from './types';
+import { FETCH_CATEGORIES, FETCH_PRODUCTS } from './types';
 import apiData from '../../apiHolder';
 import { categories, products } from '../../graphql/queries';
 
@@ -8,6 +8,15 @@ export const fetchCategories = () => async (dispatch) => {
 
   dispatch({
     type: FETCH_CATEGORIES,
+    payload: data.categories,
+  });
+};
+
+export const fetchProducts = () => async (dispatch) => {
+  const { data } = await apiData(products);
+
+  dispatch({
+    type: FETCH_PRODUCTS,
     payload: data.categories,
   });
 };
