--- src/redux/reducers/cartReducers.js
@@ -1,13 +1,0 @@
-/* eslint-disable import/no-anonymous-default-export */
-import { ADD_TO_CART } from '../action/types';
-
-const initialState = { items: [] };
-export default function reducer(state = initialState, action) {
-  switch (action.type) {
-    case ADD_TO_CART:
-      return { ...state, items: action.payload.cartItems };
-
-    default:
-      return state;
-  }
-}
