+++ src/redux/reducers/categoriesReducer.js
